refactor: Rename TaskSerializer to StorageUtils
